requirements.txt
Flask==2.3.3
pandas==2.1.1
plotly==5.17.0
openai==1.3.5
openpyxl==3.1.2
python-dotenv==1.0.0 